apply namespacing to M3 service objects
