+++ spec/services/m3/profile_importer_spec.rb
@@ -1,0 +1,23 @@
+require 'rails_helper'
+
+RSpec.describe M3::ProfileImporter do
+
+  describe 'class attributes' do
+    it '#default_logger' do
+      M3::ProfileImporter.default_logger == Rails.logger
+    end
+
+    it '#default_config_file' do
+      path_to_file = M3::ProfileImporter.default_config_file
+      expect(path_to_file).to include('config/metadata_profiles/')
+    end
+  end
+
+  describe '#load_profile_from_path' do
+    it 'returns an M3::Profile instance' do
+      expect(M3::ProfileImporter.load_profile_from_path)
+        .to be_an_instance_of(M3::Profile)
+    end
+  end
+
+end
